Update Angular SPA
