--- src/ServiceStackVS/ProjectTemplates/AngularSPA/AngularSPA/bower.json
@@ -1,25 +1,0 @@
-{
-  "name": "$safeprojectname$",
-  "version": "0.0.1",
-  "authors": [
-    ""
-  ],
-  "license": "MIT",
-  "private": true,
-  "dependencies": {
-    "angular": "~1.2.23",
-    "angular-route": "1.2.x",
-    "angular-loader": "1.2.x",
-    "angular-mocks": "~1.2.x",
-    "bootstrap": "~3.2.0",
-    "modernizr": "~2.8.3",
-    "jquery": "~1.11.3" 
-  },
-  "ignore": [
-    "**/.*",
-    "node_modules",
-    "bower_components",
-    "test",
-    "tests"
-  ]
-}
