+++ src/ServiceStackVS/ProjectTemplates/AngularSPA/AngularSPA/package.json
@@ -7,29 +7,32 @@
     "private": true,
     "license": "",
     "repository": { },
-  "devDependencies": {
-    "karma": "~0.12",
-    "karma-jasmine": "~0.2.0",
-    "karma-chrome-launcher": "~0.1.4",
-    "karma-phantomjs-launcher": "~0.1.4",
-    "karma-junit-reporter": "^0.2.2",
-    "grunt-karma": "~0.8.3",
-	"grunt-gulp":"~0.1.0",
-    "grunt": "~0.4.5",
-    "ssvs-utils": "~0.1.1",
-    "grunt-msbuild": "~0.2.2",
-    "grunt-nuget":  "0.1.3",
-    "gulp": "~3.8.7",
-    "gulp-replace": "~0.4.0",
-    "gulp-htmlbuild": "~0.4.1",
-    "event-stream":  "~3.3.1",
+  "dependencies": {
+    "angular": "~1.5.0",
+    "angular-route": "~1.5.0",
+    "angular-loader": "~1.5.0",
+    "angular-mocks": "~1.5.0",
+    "bootstrap": "~3.2.0",
+    "modernizr": "~3.3.1",
+    "jquery": "~1.11.3" 
+  },
+    "devDependencies": {
     "del": "~1.1.1",
-    "gulp-uglify": "~1.0.1",
+    "event-stream": "~3.3.1",
+    "gulp": "~3.9.1",
+    "gulp-htmlbuild": "~0.4.1",
     "gulp-if": "~1.2.4",
-    "gulp-grunt": "~0.5.2",
     "gulp-minify-css": "~0.3.8",
-    "gulp-useref": "~1.0.1",
-    "gulp-newer": "~0.3.0",
-    "requirejs":  "~2.1.15"
+    "gulp-msbuild": "^0.2.21",
+    "gulp-msdeploy": "^0.0.4",
+    "gulp-newer": "~1.1.0",
+    "gulp-nuget-restore": "^0.3.1",
+    "gulp-rename": "^1.2.2",
+    "gulp-replace": "~0.4.0",
+    "gulp-uglify": "~1.5.3",
+    "gulp-useref": "~3.0.7",
+    "requirejs": "~2.1.15",
+    "run-sequence": "^1.1.5",
+    "yargs": "^4.2.0"
   }
 }
