+++ src/ServiceStackVS/ProjectTemplates/AngularSPA/AngularSPA/wwwroot_build/00-install-dependencies.bat
@@ -2,4 +2,3 @@ REM pre-requisite that only needs to run once for installing npm and bower depen
 
 cd ..
 npm install
-bower install
